refactor(model): future_event is common
