+++ scripts/models/patrol/patrol_schema_item.py
@@ -6,6 +6,7 @@ from typing import Dict, List, Literal, Optional, Tuple, Union
 from pydantic import BaseModel, Field, ConfigDict
 from pydantic_core import MISSING
 from scripts.models.common.biome import Biome
+from scripts.models.common.future_event import FutureEvent
 from scripts.models.common.min_max_status import MinMaxStatusDictKey
 from scripts.models.common.relationship_status import RelationshipStatus
 from scripts.models.common.season import Season
@@ -79,3 +80,4 @@ class PatrolSchemaItem(BaseModel):
     fail_outcomes: List[Outcome]
     antag_success_outcomes: Union[List[Outcome], MISSING] = MISSING
     antag_fail_outcomes: List[Outcome] = MISSING
+    future_event: Union[FutureEvent, MISSING] = MISSING
